+++ js/abstract/file_web.js
@@ -1,0 +1,3 @@
+export function fetchVirtualFS(module) {
+    return module.FS;
+}